+++ src/ChunkSplitters.jl
@@ -4,7 +4,7 @@ include("api.jl")
 include("internals.jl")
 
 export chunk_indices, chunk
-export Split, BatchSplit, ScatterSplit
+export Split, Consecutive, RoundRobin
 
 if VERSION >= v"1.11.0-DEV.469"
     eval(Meta.parse("public is_chunkable"))